+++ src/main/java/gregapi/data/CS.java
@@ -1584,7 +1584,7 @@ public class CS {
 		
 		public static BlockBase CFoam, CFoamFresh, Concrete, ConcreteReinforced, Asphalt, Glass, GlowGlass;
 		public static BlockBase Bars_Wood, Bars_Steel, Bars_Titanium, Bars_TungstenSteel, Bars_Adamantium;
-		public static BlockBase Spikes_Sharp, Spikes_Super, Spikes_Metal, Spikes_Fancy;
+		public static BlockBase Spikes_Sharp, Spikes_Steel, Spikes_Super, Spikes_Metal, Spikes_Fancy;
 		public static BlockBase LongDistWire01, LongDistPipe01;
 		
 		public static IBlockBase